✨ [feat]add img upload failed prompt
